requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.4
django-browser-reload==1.12.1
django-celery-results==2.5.1
django-colorfield==0.11.0
django-filter==24.2
django-formset==1.4.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-semantic-admin==0.5.0
django-semantic-forms==0.1.5
django-treenode==0.22.0
django-unfold==0.22.0
fakeredis==2.22.0
kombu==5.3.7
pillow==10.3.0
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
rules==3.3
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
